tugasakhir5/requirements.txt
praw
pymongo
kafka-python
gensim
numpy
datetime
nltk
pandas
scipy==1.12